IssueNo:  #I3EPPI Description: platform directory refactoring Sig: kernel Feature or Bugfix: Feature Binary Source: No

+++ arch/arm/arm/src/los_hw_tick.c
@@ -65,9 +65,9 @@ LITE_OS_SEC_TEXT_MINOR VOID LOS_GetCpuCycle(UINT32 *highCnt, UINT32 *lowCnt)
 
 LITE_OS_SEC_TEXT_MINOR UINT64 LOS_CurrNanosec(VOID)
 {
-    DOUBLE nanos;
+    UINT64 nanos;
 
-    nanos = (DOUBLE)HalClockGetCycles() * OS_SYS_NS_PER_SECOND / g_sysClock;
+    nanos = (UINT64)HalClockGetCycles() * OS_SYS_NS_PER_SECOND / g_sysClock;
     return (UINT64)nanos;
 }
